pyproject.toml
[project]
name = "pdf-invoke"
version = "0.1.0"
description = "A LLM utility for working with pdfs"
authors = [
    {name = "Luciano Bermudez",email = "lberm007@ucr.edu"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "langchain-core (>=1.2.11,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pymupdf (>=1.27.1,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
